admin/requirements.txt
Django==4.0.4
django-admin-search==0.3.15
django-filter==21.1
psycopg2==2.9.5
orjson==3.8.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
django-admin-multi-select-filter==1.4.1
ipython==8.2.0
redis==4.3.5
django-tinymce==3.5.0
django-admin-rangefilter==0.8.4
markdown2==2.4.3
django-admin-search==0.3.15
requests==2.27.1
